+++ spec/schleuder/integration/dash_addresses_spec.rb
@@ -14,27 +14,23 @@ describe 'someone sends an email to a listname-dash-address' do
 
     mail = Mail::TestMailer.deliveries.first
     Mail::TestMailer.deliveries.clear
-    output = nil
-    begin
-      output = Schleuder::Runner.new().run(mail.to_s, list.sendkey_address)
-    rescue SystemExit
-    end
 
-    # if properly exited there was no output
+    output = process_mail(mail.to_s, list.sendkey_address)
     expect(output).to be_nil
 
-    raw = Mail::TestMailer.deliveries.first
-    message = raw.setup(list.email, list)
+    message = Mail::TestMailer.deliveries.first
 
     expect(message.to).to eql(['outside@example.org'])
-    expect(message.parts.first.body.to_s).to eql('Find the key for this address attached.')
-    expect(message.attachments.first.body.to_s).to include(list.fingerprint)
-    expect(message.attachments.first.body.to_s).to include('-----BEGIN PGP PUBLIC KEY BLOCK-----')
+    signed_message_parts = message.parts[0].parts
+    expect(signed_message_parts.first.body.to_s).to eql('Find the key for this address attached.')
+    expect(message.parts[0].attachments.first.body.to_s).to include(list.fingerprint)
+    expect(message.parts[0].attachments.first.body.to_s).to include('-----BEGIN PGP PUBLIC KEY BLOCK-----')
     expect(message.in_reply_to).to eql(mail.message_id)
   end
 
   it "forwards the message to the admins if extension is -owner" do
     list = create(:list)
+    # owner needs a key so they get email
     list.subscribe("schleuder@example.org", '59C71FB38AEE22E091C78259D06350440F759BD3', true)
     list.subscribe("admin@example.org", '59C71FB38AEE22E091C78259D06350440F759BD3', true)
     ENV['GNUPGHOME'] = list.listdir
@@ -47,19 +43,17 @@ describe 'someone sends an email to a listname-dash-address' do
 
     mail = Mail::TestMailer.deliveries.first
     Mail::TestMailer.deliveries.clear
-    output = nil
-    begin
-      output = Schleuder::Runner.new().run(mail.to_s, list.owner_address)
-    rescue SystemExit
-    end
 
-    # if properly exited there was no output
+    output = process_mail(mail.to_s, list.owner_address)
     expect(output).to be_nil
 
     raw_msgs = Mail::TestMailer.deliveries
     raw_msgs.sort_by { |msg| msg.to.first }
-    message1 = raw_msgs[0].setup(list.email, list)
-    message2 = raw_msgs[1].setup(list.email, list)
+    # reparse the messages so we decrypt and remove all the craft
+    # for easier parsing afterwards
+    message1, message2 = raw_msgs[0..1].collect{|m|
+      Mail.create_message_to_list(m.to_s, list.email, list).setup
+    }
 
     expect(message1.to).to eql(['admin@example.org'])
     expect(message1.subject).to eql('help')
@@ -76,7 +70,7 @@ describe 'someone sends an email to a listname-dash-address' do
 
   it "forwards the message to the admins if extension is -bounce" do
     list = create(:list)
-    list.subscribe("admin@example.org", '59C71FB38AEE22E091C78259D06350440F759BD3', true)
+    list.subscribe("admin@example.org", nil, true)
     ENV['GNUPGHOME'] = list.listdir
     mail = Mail.new
     mail.to = list.bounce_address
@@ -87,31 +81,26 @@ describe 'someone sends an email to a listname-dash-address' do
 
     mail = Mail::TestMailer.deliveries.first
     Mail::TestMailer.deliveries.clear
-    output = nil
-    begin
-      output = Schleuder::Runner.new().run(mail.to_s, list.bounce_address).inspect
-    rescue SystemExit
-    end
 
-    # if properly exited there was no output
+    output = process_mail(mail.to_s, list.bounce_address)
     expect(output).to be_nil
 
-    raw_msg = Mail::TestMailer.deliveries.first
-    message = raw_msg.setup(list.email, list)
+    message = Mail::TestMailer.deliveries.first
 
     expect(message.to).to eql(['admin@example.org'])
     expect(message.subject).to eql(I18n.t('automated_message_subject'))
-    expect(message.parts.first.body.to_s).to eql(I18n.t('forward_automated_message_to_admins'))
-    expect(message.parts.last.mime_type).to eql('message/rfc822')
-    expect(message.parts.last.body.to_s).to include('From: mailer-daemon@example.org')
-    expect(message.parts.last.body.to_s).to include(mail.message_id)
-    expect(message.parts.last.body.to_s).to include("Subject: something")
-    expect(message.parts.last.body.to_s).to include("delivery failure")
+    signed_message_parts = message.parts[0].parts
+    expect(signed_message_parts.first.body.to_s).to eql(I18n.t('forward_automated_message_to_admins'))
+    expect(signed_message_parts.last.mime_type).to eql('message/rfc822')
+    expect(signed_message_parts.last.body.to_s).to include('From: mailer-daemon@example.org')
+    expect(signed_message_parts.last.body.to_s).to include(mail.message_id)
+    expect(signed_message_parts.last.body.to_s).to include("Subject: something")
+    expect(signed_message_parts.last.body.to_s).to include("delivery failure")
   end
 
   it "forwards the message to the admins if extension is -bounce and it's a real bounce mail" do
     list = create(:list)
-    list.subscribe("admin@example.org", '59C71FB38AEE22E091C78259D06350440F759BD3', true)
+    list.subscribe("admin@example.org", nil, true)
     ENV['GNUPGHOME'] = list.listdir
     mail = Mail.new(File.read('spec/fixtures/mails/bounce.eml'))
     mail.to = list.owner_address
@@ -119,26 +108,21 @@ describe 'someone sends an email to a listname-dash-address' do
 
     mail = Mail::TestMailer.deliveries.first
     Mail::TestMailer.deliveries.clear
-    output = nil
-    begin
-      output = Schleuder::Runner.new().run(mail.to_s, list.bounce_address)
-    rescue SystemExit
-    end
 
-    # if properly exited there was no output
+    output = process_mail(mail.to_s, list.bounce_address)
     expect(output).to be_nil
 
-    raw_msg = Mail::TestMailer.deliveries.first
-    message = raw_msg.setup(list.email, list)
+    message = Mail::TestMailer.deliveries.first
 
     expect(message.to).to eql(['admin@example.org'])
     expect(message.subject).to eql(I18n.t('automated_message_subject'))
-    expect(message.parts.first.body.to_s).to eql(I18n.t('forward_automated_message_to_admins'))
-    expect(message.parts.last.mime_type).to eql('message/rfc822')
-    expect(message.parts.last.body.to_s).to include('From: mailer-daemon@example.org')
-    expect(message.parts.last.body.to_s).to include(mail.message_id)
-    expect(message.parts.last.body.to_s).to include("Subject: something")
-    expect(message.parts.last.body.to_s).to include("delivery failure")
+    signed_message_parts = message.parts[0].parts
+    expect(signed_message_parts.first.body.to_s).to eql(I18n.t('forward_automated_message_to_admins'))
+    expect(signed_message_parts.last.mime_type).to eql('message/rfc822')
+    expect(signed_message_parts.last.body.to_s).to include('Mailer-Daemon@schleuder.example.org')
+    expect(signed_message_parts.last.body.to_s).to include(mail.message_id)
+    expect(signed_message_parts.last.body.to_s).to include("Subject: bounce test")
+    expect(signed_message_parts.last.body.to_s).to include("mailbox is full")
   end
 end