+++ spec/spec_helper.rb
@@ -47,6 +47,10 @@ RSpec.configure do |config|
     mocks.verify_partial_doubles = true
   end
 
+  Mail.defaults do
+    delivery_method :test
+  end
+
   def cleanup_gnupg_home
     ENV["GNUPGHOME"] = nil
     FileUtils.rm_rf Schleuder::Conf.lists_dir
@@ -95,10 +99,15 @@ RSpec.configure do |config|
     `SCHLEUDER_ENV=test SCHLEUDER_CONFIG=spec/schleuder.yml bin/schleuder #{command} 2>&1`
   end
 
-  Mail.defaults do
-    delivery_method :test
+  def process_mail(msg, recipient)
+    output = nil
+    begin
+      output = Schleuder::Runner.new.run(msg, recipient)
+    rescue SystemExit
+    end
+    output
   end
-  
+
   def teardown_list_and_mailer(list)
     FileUtils.rm_rf(list.listdir)
     Mail::TestMailer.deliveries.clear